// src/app/common/interfaces/user.interface.ts
export interface IUserInformation {
    userId: string;
    names: string;
    lastName: string;
    email: string;
    mobileNumber: string;
    homeAddress: string;
    dateEnrollment: string;
    documentType: string;
    documentNumber: string;
    hashCode: string;
    customerId:string;
  }
  
  export interface IUserHashInformation {
    userId: string;
    names: string;
    lastName: string;
    email: string;
    mobileNumber: string;
    homeAddress: string;
    dateEnrollment: string;
    documentType: string;
    documentNumber: string;
    hashCode: string;
    customerId: string;
  }
  